Move /send and /react modules into control/ subdir
