+++ src/controllers/dev/control/send.command.ts
@@ -8,8 +8,8 @@ import {
 import {
   RoleLevel,
   checkPrivilege,
-} from "../../middleware/privilege.middleware";
-import { CommandBuilder } from "../../types/command.types";
+} from "../../../middleware/privilege.middleware";
+import { CommandBuilder } from "../../../types/command.types";
 
 const devSend = new CommandBuilder();
